Changes some integer fields to use bigint fields instead.

+++ apps/knot/lib/knot/repo/timestamp.ex
@@ -1,8 +1,8 @@
 defmodule Knot.Repo.Type.Timestamp do
   @behaviour Ecto.Type
 
-  @spec type :: :integer
-  def type, do: :integer
+  @spec type :: :bigint
+  def type, do: :bigint
 
   @spec cast(Knot.Block.timestamp) :: {:ok, integer}
   def cast(val) when is_integer(val), do: {:ok, val}
