+++ apps/knot/priv/repo/migrations/20170912160110_create_blocks.exs
@@ -4,12 +4,12 @@ defmodule Knot.Repo.Migrations.CreateBlocks do
   def change do
     create table(:blocks, primary_key: false) do
       add :hash,            :string,  primary_key: true
-      add :height,          :integer, null: false
-      add :timestamp,       :integer, null: false
+      add :height,          :bigint,  null: false
+      add :timestamp,       :bigint,  null: false
       add :parent_hash,     :string,  null: false, size: 64
       add :content_hash,    :string,  null: false, size: 64
       add :component_hash,  :string,  null: false, size: 64
-      add :nonce,           :integer, null: false
+      add :nonce,           :bigint,  null: false
     end
 
     create index :blocks, [:hash], unique: true
